Consolidate to top-level
