+++ neuro_san/__init__.py
@@ -1,5 +1,3 @@
-
-
 # Copyright © 2023-2026 Cognizant Technology Solutions Corp, www.cognizant.com.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,10 +14,42 @@
 #
 # END COPYRIGHT
 
+from typing import Any
+from typing import Type
+
 from leaf_common.config.file_of_class import FileOfClass
+
+from neuro_san.internals.utils.deprecation_redirect import DeprecationRedirect
 
 # Normally we don't use __init__.py files to define anything,
 # but here we define some constants that point to important directories in the distribution.
 TOP_LEVEL_DIR = FileOfClass(__file__)
 DEPLOY_DIR = FileOfClass(__file__, path_to_basis="./deploy")
 REGISTRIES_DIR = FileOfClass(__file__, path_to_basis="./registries")
+
+
+_DEPRECATION_REDIRECT = DeprecationRedirect(
+    __name__,
+    # A map from old class name to new class name for compatibility
+    {
+        "internals.messages.chat_message_type.ChatMessageType":
+            "neuro_san.message.types.chat_message_type.ChatMessageType",
+        "internals.messages.origination.Origination":
+            "neuro_san.message.utils.origination.Origination",
+        "internals.parsers.structure.json_structure_parser.JsonStructureParser":
+            "neuro_san.message.parsers.structure.json_structure_parser.JsonStructureParser",
+        "message_processors.message_processor.MessageProcessor":
+            "neuro_san.message.processors.message_processor.MessageProcessor",
+        "message_processors.basic_message_processor.BasicMessageProcessor":
+            "neuro_san.message.processors.basic_message_processor.BasicMessageProcessor",
+    }
+)
+
+
+def __getattr__(old_class: str) -> Type[Any]:
+    """
+    Redirect deprecated classes
+    :param old_class: The old class name
+    :return: The redirected class
+    """
+    return _DEPRECATION_REDIRECT.redirect_class(old_class)
