--- neuro_san/internals/parsers/__init__.py
@@ -1,22 +1,0 @@
-from typing import Any
-from typing import Type
-
-from neuro_san.internals.utils.deprecation_redirect import DeprecationRedirect
-
-_DEPRECATION_REDIRECT = DeprecationRedirect(
-    __name__,
-    # A map from old class name to new class name for compatibility
-    {
-        "structure.json_structure_parser.JsonStructureParser":
-            "neuro_san.message.parsers.structure.json_structure_parser.JsonStructureParser",
-    }
-)
-
-
-def __getattr__(old_class: str) -> Type[Any]:
-    """
-    Redirect deprecated classes
-    :param old_class: The old class name
-    :return: The redirected class
-    """
-    return _DEPRECATION_REDIRECT.redirect_class(old_class)
